+++ PDKit/Protocols/UI/Alerts/AlertDispatching.swift
@@ -12,11 +12,6 @@ public protocol AlertDispatching {
 
 	var factory: AlertProducing {get set}
 
-	func presentHormoneActions(
-		at row: Index,
-		reload: @escaping () -> Void,
-		nav: @escaping () -> Void
-	)
 	func presentPillActions(for pill: Swallowable, handlers: PillCellActionHandling)
 	func presentQuantityMutationAlert(
 		oldQuantity: Int, newQuantity: Int, handlers: QuantityMutationAlertActionHandling